# Optimize query for latest pnl tick at specific time.